# src/handler/openapigenmodel/go/model_customer.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Customer API
 *
 * API for customer management
 *
 * API version: 1.0.0
 */

package openapi


import (
	"time"
)



type Customer struct {

	Id string `json:"id,omitempty"`

	Name string `json:"name,omitempty"`

	Address string `json:"address,omitempty"`

	Zip string `json:"zip,omitempty"`

	Phone string `json:"phone,omitempty"`

	Mktsegment string `json:"mktsegment,omitempty"`

	Nation string `json:"nation,omitempty"`

	Birthdate time.Time `json:"birthdate,omitempty"`

	Version int `json:"version,omitempty"`
}

// AssertCustomerRequired checks if the required fields are not zero-ed
func AssertCustomerRequired(obj Customer) error {
	return nil
}

// AssertCustomerConstraints checks if the values respects the defined constraints
func AssertCustomerConstraints(obj Customer) error {
	return nil
}